+++ ios/adyen_checkout/Sources/adyen_checkout/CheckoutPlatformApi.swift
@@ -206,17 +206,6 @@ class CheckoutPlatformApi: CheckoutPlatformInterface {
     }
 
     private func getViewController() -> UIViewController? {
-        var rootViewController = UIApplication.shared.adyen.mainKeyWindow?.rootViewController
-        while let presentedViewController = rootViewController?.presentedViewController {
-            let type = String(describing: type(of: presentedViewController))
-            // TODO: - We need to discuss how the SDK should react if a DropInNavigationController is already displayed
-            if type == "DropInNavigationController" {
-                return nil
-            } else {
-                rootViewController = presentedViewController
-            }
-        }
-
-        return rootViewController
+        ViewControllerProvider.topViewController(skipTypeName: "DropInNavigationController")
     }
 }